rationalised display of links when printing Previously, the print style sheet output the url of a link immediately after the link in the text (in parens), except for those in a few named areas. This produced ugly rendering and too many URLs.
The new approach requires the content author to give links that should show their URL on printing a class of 'print'. When the doc is printed, those links have a footnote reference printed after (eg. "go here[2]"). The footnote can be found in a list in the Further reading section at the bottom of the page. It contains the URL for that link text.  That list is only shown when the document is printed.

The links in the Further reading section, however, continue to produce the URL in parens immediately after the link text.  In this case, that is a better solution.